feat(schematics): support @angular/cli@6.1.1
