+++ package.json
@@ -13,28 +13,29 @@
     "copy": "./scripts/copy.sh",
     "test:schematics": "yarn linknpm fast && ./scripts/test_schematics.sh",
     "test:nx": "yarn linknpm fast && ./scripts/test_nx.sh",
-    "test": "yarn linknpm fast && ./scripts/test_nx.sh && ./scripts/test_schematics.sh",
+    "test":
+      "yarn linknpm fast && ./scripts/test_nx.sh && ./scripts/test_schematics.sh",
     "checkformat": "prettier \"{packages,e2e}/**/*.ts\" --list-different"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.6.8",
-    "@angular-devkit/core": "^0.6.1",
-    "@angular-devkit/schematics": "^0.6.1",
-    "@angular/cli": "6.0.1",
-    "@angular/common": "6.0.1",
-    "@angular/compiler": "6.0.1",
-    "@angular/compiler-cli": "6.0.1",
-    "@angular/core": "6.0.1",
-    "@angular/platform-browser": "6.0.1",
-    "@angular/platform-browser-dynamic": "6.0.1",
-    "@angular/router": "6.0.1",
-    "@angular/upgrade": "6.0.1",
+    "@angular-devkit/build-angular": "^0.7.1",
+    "@angular-devkit/core": "^0.7.1",
+    "@angular-devkit/schematics": "^0.7.1",
+    "@angular/cli": "6.1.1",
+    "@angular/common": "6.1.0",
+    "@angular/compiler": "6.1.0",
+    "@angular/compiler-cli": "6.1.0",
+    "@angular/core": "6.1.0",
+    "@angular/platform-browser": "6.1.0",
+    "@angular/platform-browser-dynamic": "6.1.0",
+    "@angular/router": "6.1.0",
+    "@angular/upgrade": "6.1.0",
     "@ngrx/effects": "6.0.1",
     "@ngrx/router-store": "6.0.1",
     "@ngrx/schematics": "6.0.1",
     "@ngrx/store": "6.0.1",
     "@ngrx/store-devtools": "6.0.1",
-    "@schematics/angular": "^0.6.1",
+    "@schematics/angular": "^0.7.1",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
@@ -52,11 +53,11 @@
     "jasmine-spec-reporter": "~4.2.1",
     "jest": "^23.4.0",
     "jest-jasmine2": "^23.4.1",
-    "karma": "~1.7.1",
+    "karma": "~2.0.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-jasmine": "~1.1.1",
     "karma-webpack": "2.0.4",
-    "ng-packagr": "3.0.0-rc.2",
+    "ng-packagr": "3.0.6",
     "npm-run-all": "4.1.2",
     "opn": "^5.3.0",
     "precise-commits": "1.0.2",
@@ -67,7 +68,7 @@
     "semver": "5.4.1",
     "strip-json-comments": "2.0.1",
     "tmp": "0.0.33",
-    "tslint": "5.9.1",
+    "tslint": "5.11.0",
     "typescript": "~2.7.2",
     "viz.js": "^1.8.1",
     "yargs": "^11.0.0",
@@ -77,9 +78,6 @@
   "author": "Victor Savkin",
   "license": "MIT",
   "jest": {
-    "modulePathIgnorePatterns": [
-      "tmp",
-      "collection/.*/files"
-    ]
+    "modulePathIgnorePatterns": ["tmp", "collection/.*/files"]
   }
 }
