+++ packages/bazel/package.json
@@ -4,17 +4,18 @@
   "description": "Nrwl Extensions for Angular: Bazel Schematics",
   "main": "index.js",
   "types": "index.d.js",
-  "peerDependencies": {},
   "author": "Victor Savkin",
   "license": "MIT",
   "schematics": "./src/collection.json",
   "dependencies": {
     "@ngrx/schematics": "6.0.1",
-    "@schematics/angular": "0.4.6",
     "app-root-path": "^2.0.1",
     "npm-run-all": "4.1.2",
     "semver": "5.4.1",
     "tmp": "0.0.33",
     "yargs-parser": "10.0.0"
+  },
+  "peerDependencies": {
+    "@schematics/angular": "0.7.1"
   }
 }
